Make measurements controllers pluggable
