+++ examples/measurement/angle_createWithMouse_nosnapping.html
@@ -73,6 +73,9 @@
         pointerLens : new PointerLens(viewer)
     })
 
+    angleMeasurementsMouseControl.snapToEdge = false;
+    angleMeasurementsMouseControl.snapToVertex = false;
+
     angleMeasurementsMouseControl.activate();
 
 </script>
