+++ examples/measurement/distance_createWithMouse_snapping.html
@@ -85,10 +85,10 @@
 <canvas id="myCanvas"></canvas>
 <div class="slideout-sidebar">
     <img class="info-icon" src="../../assets/images/measure_distance_icon.png"/>
-    <h1>DistanceMeasurementPlugin with vertex-snapping</h1>
-    <h2>Click on the model to create distance measurements</h2>
+    <h1>DistanceMeasurementPlugin with DistanceMeasurementsMouseControl and PointerLens</h1>
+    <h2>Click on the model to create distance measurements, with vertex and edge snapping</h2>
     <p>In this example, we're loading a BIM model from the file system, then creating distance measurements wherever the
-        user clicks on the model.</p>
+        user clicks on the model with the mouse.</p>
     <h3>Components Used</h3>
     <ul>
         <li>
@@ -98,6 +98,14 @@
         <li>
             <a href="../../docs/class/src/plugins/DistanceMeasurementsPlugin/DistanceMeasurementsPlugin.js~DistanceMeasurementsPlugin.html"
                target="_other">DistanceMeasurementsPlugin</a>
+        </li>
+        <li>
+            <a href="../../docs/class/src/plugins/DistanceMeasurementsPlugin/DistanceMeasurementsMouseControl.js~DistanceMeasurementsMouseControl.html"
+               target="_other">DistanceMeasurementsMouseControl</a>
+        </li>
+        <li>
+            <a href="../../docs/class/src/extras/PointerLens/PointerLens.js~PointerLens.html"
+               target="_other">PointerLens</a>
         </li>
         <li>
             <a href="../../docs/class/src/plugins/XKTLoaderPlugin/XKTLoaderPlugin.js~XKTLoaderPlugin.html"
@@ -119,7 +127,7 @@
     // Import the modules we need for this example
     //------------------------------------------------------------------------------------------------------------------
 
-    import {Viewer, XKTLoaderPlugin, ContextMenu, DistanceMeasurementsPlugin, PointerLens} from "../../dist/xeokit-sdk.es.js";
+    import {Viewer, XKTLoaderPlugin, ContextMenu, DistanceMeasurementsPlugin, DistanceMeasurementsMouseControl, PointerLens} from "../../dist/xeokit-sdk.es.js";
 
     //------------------------------------------------------------------------------------------------------------------
     // Create a Viewer and arrange the camera
@@ -159,14 +167,16 @@
     });
 
     //------------------------------------------------------------------------------------------------------------------
-    // DistanceMeasurementsPlugin
+    // DistanceMeasurementsPlugin, DistanceMeasurementsMouseControl and PointerLens
     //------------------------------------------------------------------------------------------------------------------
 
-    const pointerLens = new PointerLens(viewer);
+    const distanceMeasurements = new DistanceMeasurementsPlugin(viewer);
 
-    const distanceMeasurements = new DistanceMeasurementsPlugin(viewer, {
-        pointerLens
-    });
+    const distanceMeasurementsMouseControl  = new DistanceMeasurementsMouseControl(distanceMeasurements, {
+        pointerLens : new PointerLens(viewer)
+    })
+
+    distanceMeasurementsMouseControl.activate();
 
     //------------------------------------------------------------------------------------------------------------------
     // Create a context menu to delete and configure measurements
@@ -271,9 +281,6 @@
         distanceMeasurementsContextMenu.show(e.event.clientX, e.event.clientY);
         e.event.preventDefault();
     });
-
-    distanceMeasurements.control.activate();
-
 
     window.viewer = viewer;
