+++ examples/measurement/index.html
@@ -258,27 +258,28 @@
 
         "Measuring Distances": [
 
-            "# Models with existing distance measurements",
+            "# Creating distance measurements with mouse",
+            ["distance_createWithMouse_snapping", "Click with mouse on model to create distance measurements"],
+            ["distance_createWithMouse_snapping_Lyon", "Click with mouse on model to create distance measurements"],
 
-            ["distance_modelWithMeasurements", "BIM model with distance measurements"],
+            ["distance_createWithMouse_nosnapping", "Click with mouse on model to create distance measurements, snapping disabled"],
+
+            "# Creating distance measurements programmatically",
+            ["distance_modelWithMeasurements", "Creating distance measurements programmatically"],
             ["distance_modelWithMeasurements_hideAxisWires", "BIM model with distance measurements and axis-aligned wires hidden"],
-            ["distance_unitsAndScale", "BIM model with configured units (see source code)"],
-
-            "# Creating distance measurements",
-
-            ["distance_createWithMouse_snapping", "Click on model to create distance measurements; long-click to snap to nearest vertex or edge"],
-            ["distance_createWithMouse_snapping_Lyon", "Click on model to create distance measurements; long-click to snap to nearest vertex or edge"]
+            ["distance_unitsAndScale", "BIM model with configured units (see source code)"]
         ],
 
         "Measuring Angles": [
 
-            "# Models with existing angle measurements",
+            "# Creating angle measurements with mouse",
+            ["angle_createWithMouse_snapping", "Click with mouse on model to create angle measurements"],
 
-            ["angle_modelWithMeasurements", "IFC 2x3 BIM model with angle measurements"],
 
-            "# Creating angle measurements",
+            ["angle_createWithMouse_nosnapping", "Click with mouse on model to create angle measurements, snapping disabled"],
 
-            ["angle_createWithMouse_snapping", "Click or tap on the model to create angle measurements, long-click to snap to nearest vertex or edge"]
+            "# Creating angle measurements programmatically",
+            ["angle_modelWithMeasurements", "Creating angle measurements programmatically"]
         ]
     };
