+++ src/plugins/AngleMeasurementsPlugin/AngleMeasurementsMouseControl.js
@@ -1,0 +1,393 @@
+import {math} from "../../viewer/scene/math/math.js";
+import {AngleMeasurementsControl} from "./AngleMeasurementsControl.js";
+
+const MOUSE_FINDING_ORIGIN = 0;
+const MOUSE_FINDING_CORNER = 1;
+const MOUSE_FINDING_TARGET = 2;
+
+/**
+ * Creates {@link AngleMeasurement}s in an {@link AngleMeasurementsPlugin} from mouse input.
+ *
+ * ## Usage
+ *
+ * [[Run example](/examples/measurement/#angle_createWithMouse_snapping)]
+ *
+ * ````javascript
+ * import {Viewer, XKTLoaderPlugin, AngleMeasurementsPlugin, AngleMeasurementsMouseControl, PointerLens} from "xeokit-sdk.es.js";
+ *
+ * const viewer = new Viewer({
+ *     canvasId: "myCanvas",
+ * });
+ *
+ * viewer.camera.eye = [-3.93, 2.85, 27.01];
+ * viewer.camera.look = [4.40, 3.72, 8.89];
+ * viewer.camera.up = [-0.01, 0.99, 0.039];
+ *
+ * const xktLoader = new XKTLoaderPlugin(viewer);
+ *
+ * const sceneModel = xktLoader.load({
+ *     id: "myModel",
+ *     src: "Duplex.xkt"
+ * });
+ *
+ * const angleMeasurements = new AngleMeasurementsPlugin(viewer);
+ *
+ * const angleMeasurementsMouseControl  = new AngleMeasurementsMouseControl(angleMeasurements, {
+ *     pointerLens : new PointerLens(viewer)
+ * })
+ *
+ * angleMeasurementsMouseControl.snapToVertex = true;
+ * angleMeasurementsMouseControl.snapToEdge = true;
+ *
+ * angleMeasurementsMouseControl.activate();
+ * ````
+ */
+export class AngleMeasurementsMouseControl extends AngleMeasurementsControl {
+
+    /**
+     * Creates a AngleMeasurementsMouseControl bound to the given AngleMeasurementsPlugin.
+     */
+    constructor(angleMeasurementsPlugin, cfg = {}) {
+
+        super(angleMeasurementsPlugin.viewer.scene);
+
+        this.pointerLens = cfg.pointerLens;
+
+        this._active = false;
+        this._mouseState = MOUSE_FINDING_ORIGIN;
+
+        this._currentAngleMeasurement = null;
+
+        const markerDiv = document.createElement('div');
+        const canvas = this.scene.canvas.canvas;
+        canvas.parentNode.insertBefore(markerDiv, canvas);
+
+        markerDiv.style.background = "black";
+        markerDiv.style.border = "2px solid blue";
+        markerDiv.style.borderRadius = "10px";
+        markerDiv.style.width = "5px";
+        markerDiv.style.height = "5px";
+        markerDiv.style.margin = "-200px -200px";
+        markerDiv.style.zIndex = "100";
+        markerDiv.style.position = "absolute";
+        markerDiv.style.pointerEvents = "none";
+
+        this.markerDiv = markerDiv;
+        this._onMouseHoverSurface = null;
+        this._onHoverNothing = null;
+        this._onPickedNothing = null;
+        this._onPickedSurface = null;
+
+        this._onInputMouseDown = null;
+        this._onInputMouseUp = null;
+
+        this._snapToEdge = cfg.snapToEdge !== false;
+        this._snapToVertex = cfg.snapToVertex !== false;
+
+        this._attachPlugin(angleMeasurementsPlugin, cfg);
+    }
+
+    _attachPlugin(angleMeasurementsPlugin, cfg = {}) {
+
+        /**
+         * The {@link AngleMeasurementsPlugin} that owns this AngleMeasurementsMouseControl.
+         *
+         * @type {AngleMeasurementsPlugin}
+         */
+        this.angleMeasurementsPlugin = angleMeasurementsPlugin;
+
+        /**
+         * The {@link AngleMeasurementsPlugin} that owns this AngleMeasurementsMouseControl.
+         *
+         * @type {AngleMeasurementsPlugin}
+         */
+        this.plugin = angleMeasurementsPlugin;
+    }
+
+    /**
+     * Gets if this AngleMeasurementsMouseControl is currently active, where it is responding to input.
+     *
+     * @returns {boolean}
+     */
+    get active() {
+        return this._active;
+    }
+
+    /**
+     * Sets whether snap-to-vertex is enabled for this AngleMeasurementsMouseControl.
+     *
+     * This is `true` by default.
+     *
+     * @param snapToVertex True if this nap-to-vertex is to be enabled for this AngleMeasurementsMouseControl.
+     */
+    set snapToVertex(snapToVertex) {
+        this._snapToVertex = snapToVertex;
+    }
+
+    /**
+     * Gets whether snap-to-vertex is enabled for this AngleMeasurementsMouseControl.
+     *
+     * This is `true` by default.
+     *
+     * @returns {boolean} True if snap-to-vertex is enabled for this AngleMeasurementsMouseControl.
+     */
+    get snapToVertex() {
+        return this._snapToVertex;
+    }
+
+    /**
+     * Sets whether snap-to-edge is enabled for this AngleMeasurementsMouseControl.
+     *
+     * This is `true` by default.
+     *
+     * @param {boolean} snapToEdge True if this snap-to-edge is to be enabled for this AngleMeasurementsMouseControl.
+     */
+    set snapToEdge(snapToEdge) {
+        this._snapToEdge = snapToEdge;
+    }
+
+    /**
+     * Gets whether snap-to-edge is enabled for this AngleMeasurementsMouseControl.
+     *
+     * This is `true` by default.
+     *
+     * @returns {boolean} True if this snap-to-edge is enabled for this AngleMeasurementsMouseControl.
+     */
+    get snapToEdge() {
+        return this._snapToEdge;
+    }
+
+    /**
+     * Activates this AngleMeasurementsMouseControl, ready to respond to input.
+     */
+    activate() {
+        if (this._active) {
+            return;
+        }
+        const angleMeasurementsPlugin = this.angleMeasurementsPlugin;
+        const scene = this.scene;
+        const input = scene.input;
+        const canvas = scene.canvas.canvas;
+        const clickTolerance = 20;
+        const cameraControl = this.angleMeasurementsPlugin.viewer.cameraControl;
+        const pointerLens = this.pointerLens;
+        let mouseHovering = false;
+        let mouseHoverEntity = false;
+        let lastMouseCanvasX = 0;
+        let lastMouseCanvasY = 0;
+        const mouseWorldPos = math.vec3();
+        const mouseHoverCanvasPos = math.vec2();
+        this._currentAngleMeasurement = null;
+        this._onMouseHoverSurface = cameraControl.on("hoverSnapOrSurface", event => {
+            if (event.snappedToVertex || event.snappedToEdge) {
+                if (pointerLens) {
+                    pointerLens.visible = true;
+                    pointerLens.centerPos = event.cursorPos || event.canvasPos;
+                    pointerLens.cursorPos = event.canvasPos;
+                    pointerLens.snapped = true;
+                }
+                this.markerDiv.style.background = "greenyellow";
+                this.markerDiv.style.border = "2px solid green";
+            } else {
+                if (pointerLens) {
+                    pointerLens.visible = true;
+                    pointerLens.centerPos = event.cursorPos || event.canvasPos;
+                    pointerLens.cursorPos = event.canvasPos;
+                    pointerLens.snapped = false;
+                }
+                this.markerDiv.style.background = "pink";
+                this.markerDiv.style.border = "2px solid red";
+            }
+            mouseHovering = true;
+            mouseHoverEntity = event.entity;
+            mouseWorldPos.set(event.worldPos);
+            mouseHoverCanvasPos.set(event.canvasPos);
+            switch (this._mouseState) {
+                case MOUSE_FINDING_ORIGIN:
+                    this.markerDiv.style.marginLeft = `${event.canvasPos[0] - 5}px`;
+                    this.markerDiv.style.marginTop = `${event.canvasPos[1] - 5}px`;
+                    break;
+                case MOUSE_FINDING_CORNER:
+                    if (this._currentAngleMeasurement) {
+                        this._currentAngleMeasurement.originWireVisible = true;
+                        this._currentAngleMeasurement.targetWireVisible = false;
+                        this._currentAngleMeasurement.cornerVisible = true;
+                        this._currentAngleMeasurement.angleVisible = false;
+                        this._currentAngleMeasurement.corner.worldPos = event.worldPos;
+                    }
+                    this.markerDiv.style.marginLeft = `-10000px`;
+                    this.markerDiv.style.marginTop = `-10000px`;
+                    canvas.style.cursor = "pointer";
+                    break;
+                case MOUSE_FINDING_TARGET:
+                    if (this._currentAngleMeasurement) {
+                        this._currentAngleMeasurement.targetWireVisible = true;
+                        this._currentAngleMeasurement.targetVisible = true;
+                        this._currentAngleMeasurement.angleVisible = true;
+                        this._currentAngleMeasurement.target.worldPos = event.worldPos;
+                    }
+                    this.markerDiv.style.marginLeft = `-10000px`;
+                    this.markerDiv.style.marginTop = `-10000px`;
+                    canvas.style.cursor = "pointer";
+                    break;
+            }
+        });
+        this._onInputMouseDown = input.on("mousedown", (coords) => {
+            lastMouseCanvasX = coords[0];
+            lastMouseCanvasY = coords[1];
+        });
+        this._onInputMouseUp = input.on("mouseup", (coords) => {
+            if (coords[0] > lastMouseCanvasX + clickTolerance ||
+                coords[0] < lastMouseCanvasX - clickTolerance ||
+                coords[1] > lastMouseCanvasY + clickTolerance ||
+                coords[1] < lastMouseCanvasY - clickTolerance) {
+                return;
+            }
+            switch (this._mouseState) {
+                case MOUSE_FINDING_ORIGIN:
+                    if (mouseHovering) {
+                        this._currentAngleMeasurement = this.angleMeasurementsPlugin.createMeasurement({
+                            id: math.createUUID(),
+                            origin: {
+                                entity: mouseHoverEntity,
+                                worldPos: mouseWorldPos
+                            },
+                            corner: {
+                                entity: mouseHoverEntity,
+                                worldPos: mouseWorldPos
+                            },
+                            target: {
+                                entity: mouseHoverEntity,
+                                worldPos: mouseWorldPos
+                            },
+                            approximate: true
+                        });
+                        this._currentAngleMeasurement.clickable = false;
+                        this._currentAngleMeasurement.originVisible = true;
+                        this._currentAngleMeasurement.originWireVisible = true;
+                        this._currentAngleMeasurement.cornerVisible = false;
+                        this._currentAngleMeasurement.targetWireVisible = false;
+                        this._currentAngleMeasurement.targetVisible = false;
+                        this._currentAngleMeasurement.angleVisible = false;
+                        this._mouseState = MOUSE_FINDING_CORNER;
+                        this.angleMeasurementsPlugin.fire("measurementStart", this._currentAngleMeasurement);
+                    }
+                    break;
+                case MOUSE_FINDING_CORNER:
+                    if (mouseHovering) {
+                        this._currentAngleMeasurement.targetWireVisible = false;
+                        this._currentAngleMeasurement.targetVisible = true;
+                        this._currentAngleMeasurement.angleVisible = true;
+                        this._mouseState = MOUSE_FINDING_TARGET;
+                    } else {
+                        if (this._currentAngleMeasurement) {
+                            this._currentAngleMeasurement.destroy();
+                            this._currentAngleMeasurement = null;
+                            this._mouseState = MOUSE_FINDING_ORIGIN
+                            this.angleMeasurementsPlugin.fire("measurementCancel", this._currentAngleMeasurement);
+                        }
+                    }
+                    break;
+                case MOUSE_FINDING_TARGET:
+                    if (mouseHovering) {
+                        this._currentAngleMeasurement.targetVisible = true;
+                        this._currentAngleMeasurement.angleVisible = true;
+                        this.angleMeasurementsPlugin.fire("measurementEnd", this._currentAngleMeasurement);
+                        this._currentAngleMeasurement = null;
+                        this._mouseState = MOUSE_FINDING_ORIGIN;
+                    } else {
+                        if (this._currentAngleMeasurement) {
+                            this._currentAngleMeasurement.destroy();
+                            this._currentAngleMeasurement = null;
+                            this._mouseState = MOUSE_FINDING_ORIGIN;
+                            this.angleMeasurementsPlugin.fire("measurementCancel", this._currentAngleMeasurement);
+                        }
+                    }
+                    break;
+            }
+        });
+        this._onMouseHoverOff = cameraControl.on("hoverSnapOrSurfaceOff", event => {
+            mouseHovering = false;
+            if (pointerLens) {
+                pointerLens.visible = true;
+                pointerLens.centerPos = event.cursorPos;
+                pointerLens.cursorPos = event.cursorPos;
+                pointerLens.snapped = false;
+            }
+            this.markerDiv.style.marginLeft = `-100px`;
+            this.markerDiv.style.marginTop = `-100px`;
+            if (this._currentAngleMeasurement) {
+                switch (this._mouseState) {
+                    case MOUSE_FINDING_ORIGIN:
+                        this._currentAngleMeasurement.originVisible = false;
+                        break;
+                    case MOUSE_FINDING_CORNER:
+                        this._currentAngleMeasurement.cornerVisible = false;
+                        this._currentAngleMeasurement.originWireVisible = false;
+                        this._currentAngleMeasurement.targetVisible = false;
+                        this._currentAngleMeasurement.targetWireVisible = false;
+                        this._currentAngleMeasurement.angleVisible = false;
+                        break;
+                    case MOUSE_FINDING_TARGET:
+                        this._currentAngleMeasurement.targetVisible = false;
+                        this._currentAngleMeasurement.targetWireVisible = false;
+                        this._currentAngleMeasurement.angleVisible = false;
+                        break;
+                }
+                canvas.style.cursor = "default";
+            }
+        });
+        this._active = true;
+    }
+
+    /**
+     * Deactivates this AngleMeasurementsMouseControl, making it unresponsive to input.
+     *
+     * Destroys any {@link AngleMeasurement} under construction by this AngleMeasurementsMouseControl.
+     */
+    deactivate() {
+        if (!this._active) {
+            return;
+        }
+        if (this.pointerLens) {
+            this.pointerLens.visible = false;
+        }
+        this.reset();
+        const input = this.angleMeasurementsPlugin.viewer.scene.input;
+        const cameraControl = this.angleMeasurementsPlugin.viewer.cameraControl;
+        input.off(this._onInputMouseDown);
+        input.off(this._onInputMouseUp);
+        cameraControl.off(this._onMouseHoverSurface);
+        cameraControl.off(this._onPickedSurface);
+        cameraControl.off(this._onHoverNothing);
+        cameraControl.off(this._onPickedNothing);
+        this._currentAngleMeasurement = null;
+        this._active = false;
+    }
+
+    /**
+     * Resets this AngleMeasurementsMouseControl.
+     *
+     * Destroys any {@link AngleMeasurement} under construction by this AngleMeasurementsMouseControl.
+     *
+     * Does nothing if the AngleMeasurementsMouseControl is not active.
+     */
+    reset() {
+        if (!this._active) {
+            return;
+        }
+        if (this._currentAngleMeasurement) {
+            this._currentAngleMeasurement.destroy();
+            this._currentAngleMeasurement = null;
+        }
+        this._mouseState = MOUSE_FINDING_ORIGIN;
+    }
+
+    /**
+     * Destroys this DistanceMeasurementsMouseControl.
+     */
+    destroy() {
+        this.deactivate();
+        super.destroy();
+    }
+}
