+++ src/plugins/AngleMeasurementsPlugin/index.js
@@ -1,1 +1,3 @@
 export * from "./AngleMeasurementsPlugin.js";
+export * from "./AngleMeasurementsControl.js";
+export * from "./AngleMeasurementsMouseControl.js";
