+++ src/plugins/DistanceMeasurementsPlugin/index.js
@@ -1,1 +1,3 @@
 export * from "./DistanceMeasurementsPlugin.js";
+export * from "./DistanceMeasurementsControl.js";
+export * from "./DistanceMeasurementsMouseControl.js";
