+++ src/viewer/scene/scene/Scene.js
@@ -2216,15 +2216,15 @@ class Scene extends Component {
      * @param {Object} params Picking parameters.
      * @param {Number[]} [params.canvasPos] Canvas-space coordinates. When ray-picking, this will override the **origin** and ** direction** parameters and will cause the ray to be fired through the canvas at this position, directly along the negative View-space Z-axis.
      * @param {Number} [params.snapRadius=30] The snap radius, in canvas pixels
-     * @param {boolean} [params.snapVertex=true] Whether to snap to vertex.
-     * @param {boolean} [params.snapEdge=true] Whether to snap to edge.
+     * @param {boolean} [params.snapToVertex=true] Whether to snap to vertex.
+     * @param {boolean} [params.snapToEdge=true] Whether to snap to edge.
      */
     snapPick(params) {
         return this._renderer.snapPick(
             params.canvasPos,
             params.snapRadius || 30,
-            params.snapVertex,
-            params.snapEdge,
+            params.snapToVertex,
+            params.snapToEdge,
         );
     }
